Move a lot of structs into types module, clean up dumb methods/functions
