+++ src/gateway/gateway_task.rs
@@ -5,7 +5,7 @@ use tokio::sync::Mutex;
 use super::GatewayClient;
 
 /// Handles all messages a client sends to the gateway post-handshake.
-pub(super) async fn gateway_task(connection: Arc<Mutex<super::Connection>>) {
+pub(super) async fn gateway_task(connection: Arc<Mutex<super::WebSocketConnection>>) {
     // TODO
     todo!()
 }
